check

+++ pkg/agent/hoststatus/dhcp.go
@@ -57,9 +57,16 @@ func shouldRetry(err error) bool {
 }
 
 func (c *hostStatusController) handleDHCPAdd(client dhcptypes.ClientInfo) error {
+
 	name := formatHostStatusName(c.config.ClusterAgentName, client.IP)
 	log.Logger.Debugf("Processing DHCP add event - IP: %s, MAC: %s, Active: %v, Lease: %s -> %s",
 		client.IP, client.MAC, client.Active, client.StartTime, client.EndTime)
+
+	if c.config.AgentObjSpec.Feature.DhcpServerConfig != nil && c.config.AgentObjSpec.Feature.DhcpServerConfig.EnableDhcpDiscovery == false {
+		log.Logger.Warnf("DhcpDiscovery is disabled, so ignore DHCP add event - IP: %s, MAC: %s, Active: %v, Lease: %s -> %s",
+			client.IP, client.MAC, client.Active, client.StartTime, client.EndTime)
+		return nil
+	}
 
 	// Try to get existing HostStatus
 	existing := &bmcv1beta1.HostStatus{}
@@ -146,6 +153,12 @@ func (c *hostStatusController) handleDHCPAdd(client dhcptypes.ClientInfo) error 
 func (c *hostStatusController) handleDHCPDelete(client dhcptypes.ClientInfo) error {
 	name := formatHostStatusName(c.config.ClusterAgentName, client.IP)
 	log.Logger.Debugf("Processing DHCP delete event - IP: %s, MAC: %s", client.IP, client.MAC)
+
+	if c.config.AgentObjSpec.Feature.DhcpServerConfig != nil && c.config.AgentObjSpec.Feature.DhcpServerConfig.EnableDhcpDiscovery == false {
+		log.Logger.Warnf("DhcpDiscovery is disabled, so ignore DHCP delete event - IP: %s, MAC: %s, Active: %v, Lease: %s -> %s",
+			client.IP, client.MAC, client.Active, client.StartTime, client.EndTime)
+		return nil
+	}
 
 	if err := c.client.HostStatuses().Delete(context.Background(), name, metav1.DeleteOptions{}); err != nil {
 		if errors.IsNotFound(err) {
